Finalize projects scanner
