+++ projects/abuild/test/settings_test.cpp
@@ -6,10 +6,6 @@ using atest::suite;
 using atest::test;
 
 static const auto testSuite = suite("abuild::Settings", [] {
-    test("type traits", [] {
-
-    });
-
     test("set", [] {
         abuild::BuildCache cache;
         abuild::Settings{cache};
