+++ projects/abuild/test/toolchains_test.cpp
@@ -6,10 +6,6 @@ using atest::suite;
 using atest::test;
 
 static const auto testSuite = suite("abuild::Toolchains", [] {
-    test("type traits", [] {
-
-    });
-
     test("scan", [] {
         abuild::BuildCache cache;
         abuild::Toolchains{cache};
